requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
obspy==1.4.1
gunicorn==23.0.0
